+++ opal/core/search/tests/test_extract.py
@@ -4,6 +4,7 @@ Unittests for opal.core.search.extract
 import datetime
 import json
 import os
+from pathlib import Path
 
 from django.urls import reverse
 from django.test import override_settings
@@ -17,6 +18,7 @@ from opal.tests.models import (
 )
 from opal.core.search import extract
 from six import u
+import tempfile
 
 
 MOCKING_FILE_NAME_OPEN = "opal.core.search.extract.open"
@@ -132,50 +134,55 @@ class GenerateFilesTestCase(OpalTestCase):
 
 
 class ZipArchiveTestCase(OpalTestCase):
+    @patch('opal.core.search.extract.tempfile')
+    @patch('opal.core.search.extract.generate_csv_files')
+    @patch('opal.core.search.extract.shutil.make_archive')
+    @patch('opal.core.search.extract.application.get_app')
+    def test_zip_file_writes(
+        self, get_app, make_archive, generate_csv_files, their_tempfile
+    ):
+        """
+        generate_zip_files does the followig
+        * Creates an extract directory in a temp directory
+        * Calls the function that creates/writes the csvs to that extract directory
+        * Writes all files within the extract directory to a zip directory
 
-    @patch('opal.core.search.extract.subrecords')
-    @patch('opal.core.search.extract.zipfile')
-    def test_subrecords(self, zipfile, subrecords):
-        patient, episode = self.new_patient_and_episode_please()
-        subrecords.return_value = [HatWearer, HouseOwner]
-        HatWearer.objects.create(name="Indiana", episode=episode)
-        HouseOwner.objects.create(patient=patient)
-        extract.zip_archive(models.Episode.objects.all(), 'this', self.user)
-        call_args = zipfile.ZipFile.return_value.__enter__.return_value.write.call_args_list
-        self.assertEqual(4, len(call_args))
-        self.assertTrue(call_args[0][0][0].endswith("data_dictionary.html"))
-        self.assertTrue(call_args[1][0][0].endswith("episodes.csv"))
-        self.assertTrue(call_args[2][0][0].endswith("hat_wearer.csv"))
-        self.assertTrue(call_args[3][0][0].endswith("house_owner.csv"))
+        This tests that by mocking out the creates/writes function to put a file in
+        and makes sure that get's written to a zip directory.
+        """
+        with tempfile.TemporaryDirectory() as csv_dir:
+            with tempfile.TemporaryDirectory() as zip_dir:
+                their_tempfile.TemporaryDirectory.return_value.__enter__.return_value = csv_dir
+                their_tempfile.mkdtemp.return_value = zip_dir
+                root_dir = None
 
-    @patch('opal.core.search.extract.subrecords')
-    @patch('opal.core.search.extract.zipfile')
-    def test_subrecords_if_none(self, zipfile, subrecords):
-        # if there are no subrecords we don't expect them to write to the file
-        patient, episode = self.new_patient_and_episode_please()
-        subrecords.return_value = [HatWearer, HouseOwner]
-        HouseOwner.objects.create(patient=patient)
-        extract.zip_archive(models.Episode.objects.all(), 'this', self.user)
-        call_args = zipfile.ZipFile.return_value.__enter__.return_value.write.call_args_list
-        self.assertEqual(3, len(call_args))
-        self.assertTrue(call_args[0][0][0].endswith("data_dictionary.html"))
-        self.assertTrue(call_args[1][0][0].endswith("episodes.csv"))
-        self.assertTrue(call_args[2][0][0].endswith("house_owner.csv"))
+                application = MagicMock()
+                modify_extract_fun = MagicMock()
+                application.get_modify_extract_functions.return_value = [modify_extract_fun]
+                get_app.return_value = application
 
-    @patch('opal.core.search.extract.subrecords')
-    @patch('opal.core.search.extract.zipfile')
-    def test_subrecords_if_empty_query(self, zipfile, subrecords):
-        # if there are no subrecords we don't expect them to write to the file
-        subrecords.return_value = [HatWearer, HouseOwner]
-        extract.zip_archive(models.Episode.objects.all(), 'this', self.user)
-        call_args = zipfile.ZipFile.return_value.__enter__.return_value.write.call_args_list
-        self.assertEqual(2, len(call_args))
-        self.assertTrue(call_args[0][0][0].endswith("data_dictionary.html"))
-        self.assertTrue(call_args[1][0][0].endswith("episodes.csv"))
+                def _generate_csv_files(_root_dir, episodes, user):
+                    nonlocal root_dir
+                    root_dir = _root_dir
+                generate_csv_files.side_effect = _generate_csv_files
+                episode_qs = models.Episode.objects.all()
+                make_archive.return_value = 'extract.zip'
+                result = extract.zip_archive(episode_qs, 'this', self.user)
+                make_archive_call_args = make_archive.call_args
+
+                # the directory that will contain extract.zip
+                zip_archive_dir = Path(root_dir).parent.absolute()
+                expected_zip_name = os.path.join(zip_archive_dir, 'extract')
+                make_archive_call_args.assert_called_once_with(
+                    expected_zip_name, 'zip', root_dir
+                )
+                modify_extract_fun.assert_called_once_with(
+                    episode_qs, root_dir, self.user,
+                )
+        self.assertEqual(os.path.basename(result), "extract.zip")
 
 
 class AsyncExtractTestCase(OpalTestCase):
-
     @patch('opal.core.search.tasks.extract.delay')
     def test_async(self, delay):
         extract.async_extract(self.user, 'THIS')
@@ -210,14 +217,26 @@ class TestBasicCsvRenderer(PatientEpisodeTestCase):
 
     def test_get_field_names_to_render(self):
         with patch.object(Colour, "_get_fieldnames_to_extract") as field_names:
-            field_names.return_value = ["name", "consistency_token"]
+            field_names.return_value = ["id", "name", "consistency_token"]
             renderer = extract.CsvRenderer(
-                Colour,  Colour.objects.all(), self.user
+                Colour, Colour.objects.all(), self.user
             )
             self.assertEqual(
                 renderer.fields,
-                ["name"]
+                ["id", "name"]
             )
+
+    def test_get_field_names_to_render_order(self):
+        with patch.object(Colour, "_get_fieldnames_to_extract") as field_names:
+            field_names.return_value = ["name", "id", "patient_id", "consistency_token", "episode_id"]
+            renderer = extract.CsvRenderer(
+                Colour, Colour.objects.all(), self.user
+            )
+            self.assertEqual(
+                renderer.fields,
+                ["id", "patient_id", "episode_id", "name"]
+            )
+
 
     def test_fields_uses_fields_arg(self):
             renderer = extract.CsvRenderer(
@@ -433,20 +452,9 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         self.pid_str = str(self.patient.id)
         self.eid_str = str(self.episode.id)
 
-    def test_get_header(self, field_names_to_extract):
-        field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
-        ]
-        renderer = extract.PatientSubrecordCsvRenderer(
-            PatientColour,
-            models.Episode.objects.all(),
-            self.user
-        )
-        self.assertEqual(["Episode", "Patient", "Name"], renderer.get_headers())
-
     def test_get_row(self, field_names_to_extract):
         field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
+            "id", "patient_id", "episode_id", "name", "consistency_token"
         ]
         renderer = extract.PatientSubrecordCsvRenderer(
             PatientColour,
@@ -455,12 +463,17 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         )
         rendered = renderer.get_row(self.patient_colour, self.episode.id)
         self.assertEqual(
-            [str(self.episode.id), str(self.patient.id), "blue"], rendered
+            [
+                str(self.patient_colour.id),
+                self.pid_str,
+                self.eid_str,
+                "blue"
+            ], rendered
         )
 
     def test_get_rows(self, field_names_to_extract):
         field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
+            "id", "patient_id", "episode_id", "name", "consistency_token",
         ]
         renderer = extract.PatientSubrecordCsvRenderer(
             PatientColour,
@@ -470,7 +483,7 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         rendered = list(
             renderer.get_rows()
         )
-        expected = [[self.eid_str, self.pid_str, "blue"]]
+        expected = [[str(self.patient_colour.id), self.pid_str, self.eid_str, "blue"]]
         self.assertEqual(expected, rendered)
 
     def test_get_rows_same_patient(self, field_names_to_extract):
@@ -478,7 +491,7 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         first_episode = self.patient.episode_set.first()
         last_episode = self.patient.episode_set.last()
         field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
+            "id", "patient_id", "episode_id", "name", "consistency_token"
         ]
 
         renderer = extract.PatientSubrecordCsvRenderer(
@@ -490,8 +503,18 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
             renderer.get_rows()
         )
         self.assertEqual([
-            [str(first_episode.id), str(self.patient.id), "blue"],
-            [str(last_episode.id), str(self.patient.id), "blue"]
+            [
+                str(self.patient_colour.id),
+                str(self.patient.id),
+                str(first_episode.id),
+                "blue"
+            ],
+            [
+                str(self.patient_colour.id),
+                str(self.patient.id),
+                str(last_episode.id),
+                "blue"
+            ],
         ], rendered)
 
 
@@ -505,26 +528,15 @@ class TestEpisodeSubrecordCsvRenderer(PatientEpisodeTestCase):
         self.eid_str = str(self.episode.id)
         self.pid_str = str(self.patient.id)
 
-    def test_get_header(self, field_names_to_extract):
-        field_names_to_extract.return_value = [
-            "episode_id", "name", "consistency_token", "id"
-        ]
-        renderer = extract.EpisodeSubrecordCsvRenderer(
-            Colour,
-            models.Episode.objects.all(),
-            self.user
-        )
-        self.assertEqual(["Patient", "Episode", "Name"], renderer.get_headers())
-
     def test_get_row(self, field_names_to_extract):
         field_names_to_extract.return_value = [
-            "episode_id", "name", "consistency_token", "id"
+             "id", "episode_id", "name", "consistency_token"
         ]
         renderer = extract.EpisodeSubrecordCsvRenderer(
             Colour, models.Episode.objects.all(), self.user
         )
         rendered = renderer.get_row(self.colour)
-        expected = [self.pid_str, self.eid_str, "blue"]
+        expected = [str(self.colour.id), self.pid_str, self.eid_str, "blue"]
         self.assertEqual(expected, rendered)