+++ opal/tests/test_core_application.py
@@ -129,6 +129,24 @@ class OpalApplicationTestCase(OpalTestCase):
             ["upstream.dependency"]
         )
 
+    def test_modify_extract(self):
+        modify_extract = MagicMock()
+        self.app.modify_extract = [modify_extract]
+        self.assertEqual(
+            self.app.get_modify_extract_functions(),
+            [modify_extract]
+        )
+
+    @patch('opal.core.application.module_loading.import_string')
+    def test_modify_extract_with_string(self, import_string):
+        self.app.modify_extract = ["some_fun"]
+        import_string.return_value = "actual_fun"
+        self.assertEqual(
+            self.app.get_modify_extract_functions(),
+            ["actual_fun"]
+        )
+        import_string.assert_called_once_with("some_fun")
+
 
 class GetAppTestCase(OpalTestCase):